+++ ft-sdk/src/lib.rs
@@ -20,7 +20,10 @@ pub mod from_request;
 pub mod processor;
 mod rng;
 pub mod utils;
+pub mod schema;
+pub mod session;
 
+pub use session::{SessionID, SessionData};
 pub use anyhow::{anyhow, bail, ensure, Context, Error};
 pub use auth::UserId;
 pub use crypto::{DecryptionError, EncryptedString, PlainText};