+++ CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ### ft-sdk
 
+- Decoupled session store from `auth-provider` feature.
+    `ft_sdk::session` module can be used to interact with the session store
+- Move `ft_sdk::auth::SessionID` to `ft_sdk::SessionID`. 
+    The `ft_sdk::session` replaces `ft_sdk::auth::session`
+- add `ft_sdk::SessionData` which can only be constructed through
+  `SessionId::data()` and is used to interact with the session store.
+- `ft_sdk::utils::uuid_v8` function to generate uuids.
+- Impl `Display` for `ft_sdk::PlainText`.
+- Derive `Clone` for `ft_sdk::Cookie` and `ft_sdk::Host`
+- `ft_sdk::auth::provider::user_data_by_id` convenience function to get user
+  data by id.
+- `ft_sdk::auth::ProviderData::first_email` to get the first email (verified or
+  unverified) of the user.
 - Added `ft_sdk::data::browser_redirect_with_cookie()`: redirect and set-cookie do
   not work well together, this function be used to work around that issue.